pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rnf-experiments"
version = "0.1.0"
description = "ICML 2025 — Quantization Residual Detector (RN‑F) experiments"
authors = [
  { name="Vu Anh Le", email="vu.le@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "torchvision>=0.19",
  "timm>=0.9.16",
  "transformers>=4.41",
  "bitsandbytes>=0.43",
  "tabpfn>=1.0.6",
  "datasets>=3.9",
  "peft>=0.11",
  "scikit-learn>=1.5",
  "pandas>=2.2",
  "numpy>=1.25",
  "matplotlib>=3.8",
  "tqdm",
  "pyyaml",
  "typer[all]",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "coverage",
  "ruff",
  "pre-commit",
  "ipykernel",
]

[tool.setuptools.packages.find]
where = ["src"]
